requirements.txt
Django~=2.2
Pillow
django-extensions
django-debug-toolbar
